Cleanup velocity code. removed classes that where common.
