+++ src/main/java/com/alysaa/geyserupdater/common/util/CheckBuildFile.java
@@ -2,6 +2,8 @@ package com.alysaa.geyserupdater.common.util;
 
 import com.alysaa.geyserupdater.bungee.BungeeUpdater;
 import com.alysaa.geyserupdater.spigot.SpigotUpdater;
+import com.alysaa.geyserupdater.velocity.VelocityUpdater;
+
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -45,6 +47,24 @@ public class CheckBuildFile {
         boolean exists = Files.exists(p);
         if (exists) {
             SpigotUpdater.plugin.getLogger().info("New Geyser build has been downloaded! Server restart is required!");
+            cachedResult = true;
+            return true;
+        }
+        cachedResult = false;
+        return false;
+    }
+    public static boolean checkVelocityFile(boolean forPlayer) {
+        if (forPlayer) {
+            long elapsedTime = System.currentTimeMillis() - callTime;
+            if (elapsedTime < 30 * 60 * 1000) {
+                return cachedResult;
+            }
+        }
+        callTime = System.currentTimeMillis();
+        Path p = Paths.get("plugins/GeyserUpdater/BuildUpdate/Geyser-Velocity.jar");
+        boolean exists = Files.exists(p);
+        if (exists) {
+            VelocityUpdater.logger.warn("New Geyser build has been downloaded! Velocity restart is required!");
             cachedResult = true;
             return true;
         }
