--- src/main/java/com/alysaa/geyserupdater/velocity/util/ScriptCreator.java
@@ -1,50 +1,0 @@
-package com.alysaa.geyserupdater.velocity.util;
-
-import com.alysaa.geyserupdater.velocity.VelocityUpdater;
-
-import java.io.DataOutputStream;
-import java.io.File;
-import java.io.FileOutputStream;
-import java.io.IOException;
-import java.lang.management.ManagementFactory;
-import java.util.List;
-
-public class ScriptCreator {
-
-    public static void createScript() throws IOException {
-        File file;
-        String extension;
-        if (OSUtils.isWindows()) {
-            extension = "bat";
-        } else if (OSUtils.isLinux() || OSUtils.isMac()) {
-            extension = "sh";
-        } else {
-            VelocityUpdater.logger.warn("Your OS is not supported! We support Linux, Mac, and Windows for automatic script creation!");
-            return;
-        }
-        file = new File("ServerRestartScript." + extension);
-        if (!file.exists()) {
-            FileOutputStream fos = new FileOutputStream(file);
-            DataOutputStream dos = new DataOutputStream(fos);
-            if (OSUtils.isWindows()) {
-                dos.writeBytes("@echo off\n");
-                dos.writeBytes(":restart\n");
-            } else if (OSUtils.isLinux() || OSUtils.isMac()) {
-                dos.writeBytes("#!/bin/sh\n");
-                dos.writeBytes("while true; do\n");
-            }
-            // Fetch JVM flags
-            List<String> inputArguments = ManagementFactory.getRuntimeMXBean().getInputArguments();
-            String runtimeFlags = String.join(" ", inputArguments);
-            // Write command to start server
-            dos.writeBytes("java " + runtimeFlags + " -jar " + ManagementFactory.getRuntimeMXBean().getClassPath() + " nogui\n");
-            if (OSUtils.isWindows()) {
-                dos.writeBytes("timeout 10 && Goto restart\n");
-            } else if (OSUtils.isLinux() || OSUtils.isMac()) {
-                dos.writeBytes("echo \"Server stopped, restarting in 10 seconds!\"; sleep 10; done\n");
-            }
-            VelocityUpdater.logger.warn("A custom restart script has been made for you.");
-            VelocityUpdater.logger.warn("You will need to shutdown and start the proxy with our provided restart script in order for restarts to work.");
-        }
-    }
-}
