+++ src/main/java/com/alysaa/geyserupdater/velocity/util/VelocityJoinListener.java
@@ -1,5 +1,6 @@
 package com.alysaa.geyserupdater.velocity.util;
 
+import com.alysaa.geyserupdater.common.util.CheckBuildFile;
 import com.velocitypowered.api.event.Subscribe;
 import com.velocitypowered.api.event.connection.PostLoginEvent;
 import net.kyori.adventure.text.Component;
@@ -9,7 +10,7 @@ public class VelocityJoinListener {
     @Subscribe
     public void onPostLogin(PostLoginEvent event) {
         // We allow a cached result of maximum age 30 minutes to be used
-        if (BuildFileChecker.checkVelocityFile(true)) {
+        if (CheckBuildFile.checkVelocityFile(true)) {
             if (event.getPlayer().hasPermission("gupdater.geyserupdate")) {
                 event.getPlayer().sendMessage(Component.text("[GeyserUpdater] New Geyser build has been downloaded! Velocity restart is required!"));
             }
